Move the file to private _forward module
